(Curso 1) Android parte 1: Crie a sua primeira app mobile - (Aula 04) Trabalhando com arquivo fonte do layout - (Seção 04) Criando a Activity via Android Studio: Create activity for new aluno addition and refactor activities and package names
