+++ Agenda/app/src/main/java/alura/com/br/ui/activity/FormularioAlunoActivity.java
@@ -1,0 +1,16 @@
+package alura.com.br.ui.activity;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import alura.com.br.R;
+
+public class FormularioAlunoActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_formulario_aluno);
+    }
+}
